+++ src/soopervisor/airflow/export.py
@@ -63,7 +63,7 @@ class AirflowExporter(abc.AbstractExporter):
 
     @staticmethod
     def _export(cfg, env_name, mode, until, skip_tests, skip_docker,
-                ignore_git, lazy_import):
+                ignore_git, lazy_import, task_name):
         """
         Copies the current source code to the target environment folder.
         The code along with the DAG declaration file can be copied to
@@ -74,7 +74,8 @@ class AirflowExporter(abc.AbstractExporter):
             tasks, args = commons.load_tasks(cmdr=e,
                                              name=env_name,
                                              mode=mode,
-                                             lazy_import=lazy_import)
+                                             lazy_import=lazy_import,
+                                             task_name=task_name)
 
             if not tasks:
                 raise CommanderStop(f'Loaded DAG in {mode!r} mode has no '