+++ src/soopervisor/shell/export.py
@@ -82,7 +82,7 @@ class SlurmExporter(abc.AbstractExporter):
 
     @staticmethod
     def _export(cfg, env_name, mode, until, skip_tests, skip_docker,
-                ignore_git, lazy_import):
+                ignore_git, lazy_import, task_name):
         """
         Export and submit jbs
         """
@@ -102,7 +102,8 @@ class SlurmExporter(abc.AbstractExporter):
             tasks, args = commons.load_tasks(cmdr=cmdr,
                                              name=env_name,
                                              mode=mode,
-                                             lazy_import=lazy_import)
+                                             lazy_import=lazy_import,
+                                             task_name=task_name)
 
             if not tasks:
                 raise CommanderStop(f'Loaded DAG in {mode!r} mode has no '